runtime-v1/v2: ability to disable events recording
