+++ runtime/v1/impl/src/main/java/com/walmartlabs/concord/runner/engine/ElementEventProcessor.java
@@ -20,97 +20,20 @@ package com.walmartlabs.concord.runner.engine;
  * =====
  */
 
-import com.walmartlabs.concord.client.ApiClientConfiguration;
-import com.walmartlabs.concord.client.ApiClientFactory;
-import com.walmartlabs.concord.client.ProcessEventRequest;
-import com.walmartlabs.concord.client.ProcessEventsApi;
-import io.takari.bpm.ProcessDefinitionProvider;
-import io.takari.bpm.ProcessDefinitionUtils;
 import io.takari.bpm.api.ExecutionException;
 import io.takari.bpm.model.AbstractElement;
-import io.takari.bpm.model.ProcessDefinition;
-import io.takari.bpm.model.SourceAwareProcessDefinition;
-import io.takari.bpm.model.SourceMap;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-import java.time.Instant;
-import java.time.ZoneOffset;
-import java.util.HashMap;
 import java.util.Map;
-import java.util.UUID;
 import java.util.function.Predicate;
 
-public class ElementEventProcessor {
+public interface ElementEventProcessor {
 
-    private static final Logger log = LoggerFactory.getLogger(ElementEventProcessor.class);
-
-    private final ApiClientFactory apiClientFactory;
-    private final ProcessDefinitionProvider processDefinitionProvider;
-
-    public ElementEventProcessor(ApiClientFactory apiClientFactory, ProcessDefinitionProvider processDefinitionProvider) {
-        this.apiClientFactory = apiClientFactory;
-        this.processDefinitionProvider = processDefinitionProvider;
-    }
-
-    public void process(ElementEvent event, EventParamsBuilder builder) throws ExecutionException {
-        process(event, builder, null);
-    }
-
-    public void process(ElementEvent event, EventParamsBuilder builder, Predicate<AbstractElement> filter) throws ExecutionException {
-        ProcessDefinition pd = processDefinitionProvider.getById(event.getProcessDefinitionId());
-        if (pd == null) {
-            throw new RuntimeException("Process definition not found: " + event.getProcessDefinitionId());
-        }
-
-        if (!(pd instanceof SourceAwareProcessDefinition)) {
-            return;
-        }
-
-        Map<String, SourceMap> sourceMaps = ((SourceAwareProcessDefinition) pd).getSourceMaps();
-
-        SourceMap source = sourceMaps.get(event.getElementId());
-        if (source == null) {
-            return;
-        }
-
-        AbstractElement element = ProcessDefinitionUtils.findElement(pd, event.getElementId());
-
-        if (filter != null && !filter.test(element)) {
-            return;
-        }
-
-        try {
-            Map<String, Object> e = new HashMap<>();
-            e.put("processDefinitionId", event.getProcessDefinitionId());
-            e.put("elementId", event.getElementId());
-            e.put("line", source.getLine());
-            e.put("column", source.getColumn());
-            e.put("description", source.getDescription());
-            e.putAll(builder.build(element));
-
-            ProcessEventRequest req = new ProcessEventRequest();
-            req.setEventType("ELEMENT"); // TODO should it be in the constants?
-            req.setData(e);
-            req.setEventDate(Instant.now().atOffset(ZoneOffset.UTC));
-
-            ProcessEventsApi client = new ProcessEventsApi(apiClientFactory.create(
-                    ApiClientConfiguration.builder()
-                            .sessionToken(event.getSessionToken())
-                            .txId(UUID.fromString(event.getInstanceId()))
-                            .build()));
-            client.event(UUID.fromString(event.getInstanceId()), req);
-        } catch (Exception e) {
-            log.warn("process ['{}'] -> transfer error: {}", event.getInstanceId(), e.getMessage());
-        }
-    }
-
-    public interface EventParamsBuilder {
+    interface EventParamsBuilder {
 
         Map<String, Object> build(AbstractElement element);
     }
 
-    public static class ElementEvent {
+    class ElementEvent {
 
         private final String instanceId;
         private final String processDefinitionId;
@@ -140,4 +63,8 @@ public class ElementEventProcessor {
             return sessionToken;
         }
     }
+
+    void process(ElementEvent event, EventParamsBuilder builder) throws ExecutionException;
+
+    void process(ElementEvent event, EventParamsBuilder builder, Predicate<AbstractElement> filter) throws ExecutionException;
 }
