+++ runtime/v2/model/src/main/java/com/walmartlabs/concord/runtime/v2/parser/ConfigurationGrammar.java
@@ -48,6 +48,7 @@ public final class ConfigurationGrammar {
             betweenTokens(JsonToken.START_OBJECT, JsonToken.END_OBJECT,
                     with(ImmutableEventConfiguration::builder,
                             o -> options(
+                                    optional("recordEvents", booleanVal.map(o::recordEvents)),
                                     optional("recordTaskInVars", booleanVal.map(o::recordTaskInVars)),
                                     optional("truncateInVars", booleanVal.map(o::truncateInVars)),
                                     optional("truncateMaxStringLength", intVal.map(o::truncateMaxStringLength)),
