+++ runtime/v2/model/src/test/java/com/walmartlabs/concord/project/runtime/v2/ConcordJsonSchemaGeneratorTest.java
@@ -46,6 +46,7 @@ public class ConcordJsonSchemaGeneratorTest {
 
         Set<ValidationMessage> validationResult = schema.validate(concordYml);
 
+        System.out.println(validationResult);
         assertTrue(validationResult.isEmpty());
     }
 }
