+++ runtime/v2/runner/src/main/java/com/walmartlabs/concord/runtime/v2/runner/remote/TaskCallEventRecordingListener.java
@@ -62,6 +62,10 @@ public class TaskCallEventRecordingListener implements TaskCallListener {
 
     @Override
     public void onEvent(TaskCallEvent event) {
+        if (!eventConfiguration.recordEvents()) {
+            return;
+        }
+
         Map<String, Object> m = event(event);
 
         m.put("phase", event.phase().name().toLowerCase());
